+++ CHANGELOG.md
@@ -6,3 +6,4 @@ Change log for [siemstress](https://github.com/dogoncouch/siemstress)
 - Config file
 - `--clear` option to drop table
 - `-q` option to query table
+- Add query module/CLI query tool